= move TextUtils to ui/common
